Finish the base code for tile conversion, but buggy

+++ tools/map-procedural-generation-terrain/VoronioForTiledMapTmx.h
@@ -15,7 +15,7 @@ public:
     struct PolygonZoneIndex
     {
         float area;
-        uint8_t index;
+        unsigned int index;
     };
     struct PolygonZone
     {
